fix primary restart

+++ lab-4/primary/docker-entrypoint.sh
@@ -3,12 +3,19 @@ set -e
 
 DEFAULT_PGDATA="/var/lib/postgresql/data"
 
-echo "$POSTGRES_PASSWORD" > /tmp/pgpass
-chmod 777 /tmp/pgpass
+if [ ! -f "$PGDATA/PG_VERSION" ]; then
+    echo "Инициализация базы данных PostgreSQL..."
 
-runuser -u postgres -- initdb --username="${POSTGRES_USER}" --pwfile=/tmp/pgpass -D "$PGDATA" --auth=password
+    echo "$POSTGRES_PASSWORD" > /tmp/pgpass
+    chmod 777 /tmp/pgpass
 
-rm /tmp/pgpass
-echo "Инициализация кластера завершена."
+    runuser -u postgres -- initdb --username="${POSTGRES_USER:-postgres}" --pwfile=/tmp/pgpass -D "$PGDATA" --auth=password
 
+    rm /tmp/pgpass
+    echo "Инициализация кластера завершена."
+else
+    echo "База данных уже инициализирована, пропускаем initdb."
+fi
+
+echo "Запуск PostgreSQL..."
 exec runuser -u postgres -- "$@"
